Update test.cs
Commented code to point out that this function assumes that all points are in the first quadrant. The purpose of the function is to test a point against a polygon to see if its inside or outside by comparing a ray against the count of intersections. Adjust the ray if there are points that are negative.